Update default model to one Anthropic still allows
Anthropic seems to have deprecated `claude-3-5-sonnet-20241022`
as a model. Attempts to use `efrit` with that model as a backend
result in an "unknown model" error.

Swap to `claude-3-7-sonnet-20250219` instead.

----

This feels like an arms race where we'll have to bump this forward
again in a few months. A structural change would be nice but is
beyond my current grasp of elisp.